Refactor passport command structure and implement fetch and request access functionalities
